// frontend/src/types/schema.ts
export interface MangaResponseSchema {
    result: "ok" | "error";
    /** @default "entity" */
    response: string;
    data: MangaSchema;
}

/** ChapterResponse */
export interface ChapterResponseSchema {
    result: "ok" | "error";
    /** @default "entity" */
    response: string;
    data: ChapterSchema;
}
// ...
export interface ChapterSchema {
    /** @format uuid */
    id: string;
    type: "chapter";
    attributes: ChapterAttributesSchema;
    relationships: RelationshipSchema[];
}
// ...
export type ChapterEditSchema = ChapterRequestSchema;

/** Response */
export interface ResponseSchema {
    result: "ok" | "error";
}
// ...
export interface ChapterRequestSchema {
    /** @maxLength 255 */
    title: string | null;
    volume: string | null;
    /** @maxLength 8 */
    chapter: string | null;
    /** @pattern ^[a-z]{2}(-[a-z]{2})?$ */
    translatedLanguage: string;
    /** @maxItems 10 */
    groups: string[];
    /** @min 1 */
    version: number;
}
// ...
export interface ChapterListSchema {
    /** @default "ok" */
    result: string;
    /** @default "collection" */
    response: string;
    data: ChapterSchema[];
    limit: number;
    offset: number;
    total: number;
}
// ...
export type ReferenceExpansionChapterSchema = (
    | "manga"
    | "scanlation_group"
    | "user"
)[];

/**
 * ReferenceExpansionCoverArt
 * Reference expansion options for cover art entities or lists
 */
export type ReferenceExpansionCoverArtSchema = ("manga" | "user")[];

/**
 * ReferenceExpansionManga
 * Reference expansion options for manga entities or lists
 */
export type ReferenceExpansionMangaSchema = (
    | "manga"
    | "cover_art"
    | "author"
    | "artist"
    | "tag"
    | "creator"
)[];
// ...
        export type RequestParams = {};
        export type RequestQuery = {};
        export type RequestBody = never;
        export type RequestHeaders = {};
        export type ResponseBody = string;
// ...
export namespace Manga {
// ...
    export namespace GetMangaAggregate {
        export type RequestParams = {
            /**
             * Manga ID
             * @format uuid
             */
            id: string;
        };
        export type RequestQuery = {
            translatedLanguage?: string[];
            groups?: string[];
        };
        export type RequestBody = never;
        export type RequestHeaders = {};
        export type ResponseBody = {
            /** @default "ok" */
            result?: string;
            volumes?: Record<
                string,
                {
                    volume?: string;
                    count?: number;
                    chapters?: Record<
                        string,
                        {
                            chapter?: string;
                            /** @format uuid */
                            id?: string;
                            others?: string[];
                            count?: number;
                        }
                    >;
                }
            >;
        };
    }
    /**
     * @description Get Manga.
     * @tags Manga
     * @name GetMangaId
     * @summary Get Manga
     * @request GET:/manga/{id}
     */
    export namespace GetMangaId {
        export type RequestParams = {
            /**
             * Manga ID
             * @format uuid
             */
            id: string;
        };
        export type RequestQuery = {
            /** Reference expansion options for manga entities or lists */
            includes?: ReferenceExpansionMangaSchema;
        };
        export type RequestBody = never;
        export type RequestHeaders = {};
        export type ResponseBody = MangaResponseSchema;
    }
// ...
}
// ...
export namespace Chapter {
    /**
     * @description Chapter list. If you want the Chapters of a given Manga, please check the feed endpoints.
     * @tags Chapter
     * @name GetChapter
     * @summary Chapter list
     * @request GET:/chapter
     */
    export namespace GetChapter {
        export type RequestParams = {};
        export type RequestQuery = {
            /**
             * @min 0
             * @max 100
             * @default 10
             */
            limit?: number;
            /** @min 0 */
            offset?: number;
            /** Chapter ids (limited to 100 per request) */
            ids?: string[];
            title?: string;
            groups?: string[];
            uploader?: string | string[];
            /** @format uuid */
            manga?: string;
            volume?: string | string[];
            chapter?: string | string[];
            translatedLanguage?: string[];
            originalLanguage?: string[];
            excludedOriginalLanguage?: string[];
            /** @default ["safe","suggestive","erotica"] */
            contentRating?: (
                | "safe"
                | "suggestive"
                | "erotica"
                | "pornographic"
            )[];
            excludedGroups?: string[];
            excludedUploaders?: string[];
            /** @default "1" */
            includeFutureUpdates?: "0" | "1";
            includeEmptyPages?: 0 | 1;
            includeFuturePublishAt?: 0 | 1;
            includeExternalUrl?: 0 | 1;
            /**
             * DateTime string with following format: YYYY-MM-DDTHH:MM:SS in timezone UTC+0
             * @pattern ^\d{4}-[0-1]\d-([0-2]\d|3[0-1])T([0-1]\d|2[0-3]):[0-5]\d:[0-5]\d$
             */
            createdAtSince?: string;
            /**
             * DateTime string with following format: YYYY-MM-DDTHH:MM:SS in timezone UTC+0
             * @pattern ^\d{4}-[0-1]\d-([0-2]\d|3[0-1])T([0-1]\d|2[0-3]):[0-5]\d:[0-5]\d$
             */
            updatedAtSince?: string;
            /**
             * DateTime string with following format: YYYY-MM-DDTHH:MM:SS in timezone UTC+0
             * @pattern ^\d{4}-[0-1]\d-([0-2]\d|3[0-1])T([0-1]\d|2[0-3]):[0-5]\d:[0-5]\d$
             */
            publishAtSince?: string;
            order?: {
                createdAt?: "asc" | "desc";
                updatedAt?: "asc" | "desc";
                publishAt?: "asc" | "desc";
                readableAt?: "asc" | "desc";
                volume?: "asc" | "desc";
                chapter?: "asc" | "desc";
            };
            includes?: ("manga" | "scanlation_group" | "user")[];
        };
        export type RequestBody = never;
        export type RequestHeaders = {};
        export type ResponseBody = ChapterListSchema;
    }
    /**
     * No description
     * @tags Chapter
     * @name GetChapterId
     * @summary Get Chapter
     * @request GET:/chapter/{id}
     */
    export namespace GetChapterId {
        export type RequestParams = {
            /**
             * Chapter ID
             * @format uuid
             */
            id: string;
        };
        export type RequestQuery = {
            /** Reference expansion options for chapter entities or lists */
            includes?: ReferenceExpansionChapterSchema;
        };
        export type RequestBody = never;
        export type RequestHeaders = {};
        export type ResponseBody = ChapterResponseSchema;
    }
    /**
     * No description
     * @tags Chapter
     * @name PutChapterId
     * @summary Update Chapter
     * @request PUT:/chapter/{id}
     * @secure
     */
    export namespace PutChapterId {
        export type RequestParams = {
            /**
             * Chapter ID
             * @format uuid
             */
            id: string;
        };
        export type RequestQuery = {};
        export type RequestBody = ChapterEditSchema;
        export type RequestHeaders = {
            /** @default "application/json" */
            "Content-Type": string;
        };
        export type ResponseBody = ChapterResponseSchema;
    }
    /**
     * No description
     * @tags Chapter
     * @name DeleteChapterId
     * @summary Delete Chapter
     * @request DELETE:/chapter/{id}
     * @secure
     */
    export namespace DeleteChapterId {
        export type RequestParams = {
            /**
             * Chapter ID
             * @format uuid
             */
            id: string;
        };
        export type RequestQuery = {};
        export type RequestBody = never;
        export type RequestHeaders = {};
        export type ResponseBody = ResponseSchema;
    }
}
// ...
export namespace AtHome {
    /**
     * No description
     * @tags AtHome
     * @name GetAtHomeServerChapterId
     * @summary Get MangaDex@Home server URL
     * @request GET:/at-home/server/{chapterId}
     */
    export namespace GetAtHomeServerChapterId {
        export type RequestParams = {
            /**
             * Chapter ID
             * @format uuid
             */
            chapterId: string;
        };
        export type RequestQuery = {
            /**
             * Force selecting from MangaDex@Home servers that use the standard HTTPS port 443.
             *
             * While the conventional port for HTTPS traffic is 443 and servers are encouraged to use it, it is not a hard requirement as it technically isn't
             * anything special.
             *
             * However, some misbehaving school/office network will at time block traffic to non-standard ports, and setting this flag to `true` will ensure
             * selection of a server that uses these.
             * @default false
             */
            forcePort443?: boolean;
        };
        export type RequestBody = never;
        export type RequestHeaders = {};
        export type ResponseBody = {
            /** @default "ok" */
            result?: string;
            /**
             * The base URL to construct final image URLs from.
             * The URL returned is valid for the requested chapter only, and for a duration of 15 minutes from the time of the response.
             */
            baseUrl?: string;
            chapter?: {
                hash?: string;
                data?: string[];
                dataSaver?: string[];
            };
        };
    }
}